Refactor: Separate user and seminar
