+++ seminar2/src/main/kotlin/com/wafflestudio/seminar/core/seminar/api/request/JoinSeminarRequest.kt
@@ -1,4 +1,4 @@
-package com.wafflestudio.seminar.core.user.api.request
+package com.wafflestudio.seminar.core.seminar.api.request
 
 import com.wafflestudio.seminar.core.user.domain.Role
