+++ seminar2/src/main/kotlin/com/wafflestudio/seminar/core/seminar/repository/CustomSeminarRepository.kt
@@ -1,0 +1,7 @@
+package com.wafflestudio.seminar.core.seminar.repository
+
+import com.wafflestudio.seminar.core.seminar.database.SeminarEntity
+
+interface CustomSeminarRepository {
+    fun findByNameWithOrder(name: String, order: String): List<SeminarEntity>
+}
