+++ seminar2/src/main/kotlin/com/wafflestudio/seminar/core/user/database/ParticipantProfileEntity.kt
@@ -3,7 +3,7 @@ package com.wafflestudio.seminar.core.user.database
 import com.wafflestudio.seminar.common.BaseTimeEntity
 import com.wafflestudio.seminar.core.user.domain.ParticipantProfile
 import com.wafflestudio.seminar.core.user.domain.Role
-import com.wafflestudio.seminar.core.user.domain.UserParticipantSeminar
+import com.wafflestudio.seminar.core.userSeminar.domain.UserParticipantSeminar
 import javax.persistence.*
 
 @Entity
