+++ seminar2/src/main/kotlin/com/wafflestudio/seminar/core/user/domain/InstructorProfile.kt
@@ -1,5 +1,7 @@
 package com.wafflestudio.seminar.core.user.domain
 
+import com.wafflestudio.seminar.core.userSeminar.domain.UserInstructorSeminar
+
 data class InstructorProfile(
     val id: Long,
     val company: String = "",
