+++ seminar2/src/main/kotlin/com/wafflestudio/seminar/core/user/domain/ParticipantProfile.kt
@@ -1,5 +1,7 @@
 package com.wafflestudio.seminar.core.user.domain
 
+import com.wafflestudio.seminar.core.userSeminar.domain.UserParticipantSeminar
+
 data class ParticipantProfile(
     val id: Long,
     val university: String = "",
