--- seminar2/src/main/kotlin/com/wafflestudio/seminar/core/user/repository/CustomSeminarRepository.kt
@@ -1,7 +1,0 @@
-package com.wafflestudio.seminar.core.user.repository
-
-import com.wafflestudio.seminar.core.user.database.SeminarEntity
-
-interface CustomSeminarRepository {
-    fun findByNameWithOrder(name: String, order: String): List<SeminarEntity>
-}
