--- seminar2/src/main/kotlin/com/wafflestudio/seminar/core/user/repository/CustomUserSeminarRepository.kt
@@ -1,8 +1,0 @@
-package com.wafflestudio.seminar.core.user.repository
-
-import com.wafflestudio.seminar.core.user.database.SeminarEntity
-import com.wafflestudio.seminar.core.user.database.UserSeminarEntity
-
-interface CustomUserSeminarRepository {
-    fun findByUserIdAndSeminarId(userId: Long, seminarId: Long): UserSeminarEntity?
-}
