+++ seminar2/src/main/kotlin/com/wafflestudio/seminar/core/user/service/UserService.kt
@@ -1,10 +1,11 @@
 package com.wafflestudio.seminar.core.user.service
 
+import com.wafflestudio.seminar.core.seminar.api.request.CreateSeminarRequest
+import com.wafflestudio.seminar.core.seminar.api.request.EditSeminarRequest
 import com.wafflestudio.seminar.core.user.api.request.*
-import com.wafflestudio.seminar.core.user.api.response.SeminarResponse
-import com.wafflestudio.seminar.core.user.database.UserEntity
+import com.wafflestudio.seminar.core.seminar.api.response.SeminarResponse
 import com.wafflestudio.seminar.core.user.domain.Role
-import com.wafflestudio.seminar.core.user.domain.Seminar
+import com.wafflestudio.seminar.core.seminar.domain.Seminar
 import com.wafflestudio.seminar.core.user.domain.User
 
 interface UserService {
@@ -13,10 +14,4 @@ interface UserService {
     fun getProfile(userId: Long): User
     fun editProfile(userId: Long, editProfileRequest: EditProfileRequest)
     fun registerParticipantProfile(userId: Long, participantRequest: ParticipantRequest)
-    fun createSeminar(userId: Long, createSeminarRequest: CreateSeminarRequest): Seminar
-    fun editSeminar(userId: Long, editSeminarRequest: EditSeminarRequest): Seminar
-    fun getSeminar(seminarId: Long): Seminar
-    fun getSeminars(name: String, order: String): List<SeminarResponse>
-    fun joinSeminar(userId: Long, seminarId: Long, role: Role): Seminar
-    fun dropSeminar(userId: Long, seminarId: Long): Seminar
 }
