+++ seminar2/src/main/kotlin/com/wafflestudio/seminar/core/userSeminar/repository/CustomUserSeminarRepository.kt
@@ -1,0 +1,7 @@
+package com.wafflestudio.seminar.core.userSeminar.repository
+
+import com.wafflestudio.seminar.core.userSeminar.database.UserSeminarEntity
+
+interface CustomUserSeminarRepository {
+    fun findByUserIdAndSeminarId(userId: Long, seminarId: Long): UserSeminarEntity?
+}
