+++ seminar2/src/main/kotlin/com/wafflestudio/seminar/core/userSeminar/repository/UserSeminarRepository.kt
@@ -1,6 +1,6 @@
-package com.wafflestudio.seminar.core.user.repository
+package com.wafflestudio.seminar.core.userSeminar.repository
 
-import com.wafflestudio.seminar.core.user.database.UserSeminarEntity
+import com.wafflestudio.seminar.core.userSeminar.database.UserSeminarEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface UserSeminarRepository : JpaRepository<UserSeminarEntity, Long> {
